skip already read files

+++ exps/shorten_data.py
@@ -14,10 +14,13 @@ fpaths = glob.glob(os.path.join(RES_PATH,'long', '*.json.gz'))
 print('Separating %s files..' %len(fpaths))
 for fpath in fpaths:
     fname = fpath.split('%s/long/' %RES_PATH)[1].replace('.json.gz','')
-    data = utils.read_json_compressed(fpath)
-    data_short = {k:v for k,v in data.items() if k not in ['memes', 'feeds']}
     newpath = os.path.join(RES_PATH,'%s.json' %fname)
-    fout = gzip.open(newpath, 'w')
-    utils.write_json_compressed(fout, data)
+    if utils.make_sure_file_exists(newpath):
+        continue
+    else:
+        data = utils.read_json_compressed(fpath)
+        data_short = {k:v for k,v in data.items() if k not in ['memes', 'feeds']}
+        fout = gzip.open(newpath, 'w')
+        utils.write_json_compressed(fout, data)
 
 print('Finish separating data!')
